+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [2.0.0] - 2023-08-13
+### Updated
+- The `Service` type and `NewService` function to accept a new `Config` type
+  which allows default query values, headers, user token and API keys to be
+  set.
+
 ## [1.5.0] - 2023-08-11
 ### Added
 - The `IntegrationStatus` field to the `Account` type.